portable mmap in udb-inspect. svn:NO TEST
git-svn-id: file:///svn/nsd/trunk@3305 a26ef69c-88ff-0310-839f-98b793d9c207

+++ tpkg/cutest/udb-inspect.c
@@ -393,7 +393,8 @@ inspect_file(char* fname)
 		exit(1);
 	}
 	sz = inspect_initial(fd);
-	base = mmap(NULL, sz, PROT_READ, MAP_SHARED, fd, (off_t)0);
+	base = mmap(NULL, (size_t)sz, (int)PROT_READ, (int)MAP_SHARED,
+		(int)fd, (off_t)0);
 	if(base == MAP_FAILED) {
 		printf("mmap %s %u: %s\n", fname, (unsigned)sz,
 			strerror(errno));
